shared/requirements.txt
Flask==2.3.3
PyJWT==2.8.0
psycopg2-binary==2.9.7
requests==2.31.0
python-dotenv==1.0.0
